requirement.txt
streamlit
zhipuai
requests
python-dotenv
python-docx
feedparser
pymupdf